test: consolidate migration safety invariants
